Put functions and exports in vulnValidate.js
